Revert "Feat/week 02 compose"
