--- app/src/main/java/com/sopt/now/NowSopt.kt
@@ -1,37 +1,0 @@
-package com.sopt.now
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.rememberNavController
-import com.sopt.now.compose.navigation.SetupNavigation
-import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
-
-class MainActivity : ComponentActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            NOWSOPTAndroidTheme {
-                ActivityFunc()
-            }
-        }
-    }
-}
-
-
-@Composable
-fun ActivityFunc() {
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = MaterialTheme.colorScheme.background
-    ) {
-        val navController = rememberNavController()
-        SetupNavigation(navController = navController)
-    }
-}
