--- app/src/main/java/com/sopt/now/compose/component/ContextExt.kt
@@ -1,8 +1,0 @@
-package com.sopt.now.compose.component
-
-import android.content.Context
-import android.widget.Toast
-
-fun Context.toastMessage(message: Int) {
-    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
-}
