--- app/src/main/java/com/sopt/now/compose/component/bottomnavigation/BottomNavigationItem.kt
@@ -1,8 +1,0 @@
-package com.sopt.now.compose.component.bottomnavigation
-
-import androidx.compose.ui.graphics.vector.ImageVector
-
-data class BottomNavigationItem(
-    val icon: ImageVector,
-    val label: String
-)
