--- app/src/main/java/com/sopt/now/compose/navigation/Screen.kt
@@ -1,7 +1,0 @@
-package com.sopt.now.compose.navigation
-
-sealed class Screen(val route: String) {
-    data object Login : Screen("login")
-    data object Home : Screen("home")
-    data object SignUp : Screen("signup")
-}
