--- app/src/main/java/com/sopt/now/data/model/Profile.kt
@@ -1,15 +1,0 @@
-package com.sopt.now.data.model
-
-sealed class Profile {
-    data class MyProfile(
-        val profileImage: Int,
-        val name: String,
-        val mbti: String
-    ) : Profile()
-
-    data class FriendProfile(
-        val profileImage: Int,
-        val name: String,
-        val mbti: String,
-    ) : Profile()
-}
