--- app/src/main/java/com/sopt/now/data/model/User.kt
@@ -1,9 +1,0 @@
-package com.sopt.now.data.model
-
-
-data class User(
-    val id: String,
-    val password: String,
-    val nickname: String,
-    val mbti: String
-)
