+++ app/src/main/res/values/strings.xml
@@ -1,31 +1,3 @@
 <resources>
     <string name="app_name">NOW SOPT Android</string>
-
-    <!--공통요소-->
-    <string name="all_id">ID</string>
-    <string name="all_id_hint">아이디를 입력해주세요.</string>
-    <string name="all_password">비밀번호</string>
-    <string name="all_password_hint">비밀번호를 입력해주세요.</string>
-    <string name="all_mbti">MBTI</string>
-    <string name="all_nick_name">닉네임</string>
-    <string name="all_enroll">회원가입 하기</string>
-    <string name="all_input">입력하기</string>
-
-    <!--login-->
-    <string name="login_welcome">Welcome to SOPT</string>
-    <string name="login_button">로그인 하기</string>
-    <string name="login_Success">로그인에 성공했습니다</string>
-    <!--signUp-->
-    <string name="singUp_title">SIGN UP</string>
-    <string name="singUp_nick_name_hint">닉네임를 입력해주세요.</string>
-    <string name="singUp_MBTI_hint">MBTI를 입력해주세요.</string>
-    <string name="singUp_Success">회원가입에 성공하였습니다</string>
-    <string name="singUp_failure">회원가입에 실패하였습니다</string>
-
-    <!--main-->
-    <string name="main_title">%s 님의 프로필</string>
-    <string name="main_home">홈</string>
-    <string name="main_search">검색</string>
-    <string name="main_profile">프로필</string>
-
 </resources>
